renaming transformable objects (own custom objects with a transformpanel)
