+++ 3DAmsterdam/Assets/Netherlands3D/Scripts/Interface/SelectableText.cs
@@ -11,6 +11,7 @@ namespace Netherlands3D.Interface
 {
 	public class SelectableText : MonoBehaviour, IPointerEnterHandler
 	{
+		[SerializeField]
 		private InputField inputFieldSource;
 		private InputField inputField;
 
@@ -38,10 +39,25 @@ namespace Netherlands3D.Interface
 			gameObject.AddComponent<ChangePointerStyleHandler>().StyleOnHover = ChangePointerStyleHandler.Style.TEXT;
 		}
 
-		private void FinishedInteractionWithText(string newString = "")
+		/// <summary>
+		/// Clear the generated inputfield and its listeners
+		/// </summary>
+		public void Clear()
+		{
+			if (inputField)
+			{
+				Destroy(inputField);
+				inputField.onEndEdit.RemoveAllListeners();
+			}
+		}
+
+		private void OnEndEdit(string newString = "")
 		{
 			inputField.gameObject.SetActive(false);
-
+			changedTextEvent.Invoke(newString);
+		}
+		private void OnEdit(string newString = "")
+		{
 			changedTextEvent.Invoke(newString);
 		}
 
@@ -78,7 +94,8 @@ namespace Netherlands3D.Interface
 
 			DisplayInputWithText();
 
-			inputField.onEndEdit.AddListener(FinishedInteractionWithText);
+			inputField.onEndEdit.AddListener(OnEndEdit);
+			inputField.onValueChanged.AddListener(OnEdit);
 		}
 
 		private void OnDestroy()
