+++ 3DAmsterdam/Assets/Netherlands3D/Scripts/ModelParsing/ObjLoad.cs
@@ -182,7 +182,6 @@ public class ObjLoad : MonoBehaviour
 						ObjectUsesRDCoordinates = true;
 					}
 				}
-
 				if (ObjectUsesRDCoordinates)
 				{
 					if (flipYZ)
